refactor(core): Reorganize error hierarchy in `core` and `workflow` packages (no-changelog)
