+++ packages/core/src/errors/abstract/binary-data.error.ts
@@ -1,0 +1,3 @@
+import { ApplicationError } from 'n8n-workflow';
+
+export abstract class BinaryDataError extends ApplicationError {}
