+++ packages/core/src/errors/disallowed-filepath.error.ts
@@ -1,0 +1,7 @@
+import { FileSystemError } from './abstract/filesystem.error';
+
+export class DisallowedFilepathError extends FileSystemError {
+	constructor(filePath: string) {
+		super('Disallowed path detected', filePath);
+	}
+}
