+++ packages/core/src/errors/file-not-found.error.ts
@@ -1,0 +1,7 @@
+import { FileSystemError } from './abstract/filesystem.error';
+
+export class FileNotFoundError extends FileSystemError {
+	constructor(filePath: string) {
+		super('File not found', filePath);
+	}
+}
