+++ packages/core/src/errors/index.ts
@@ -1,1 +1,5 @@
-export { BinaryFileNotFoundError, FileNotFoundError, InvalidPathError } from './filesystem.errors';
+export { FileNotFoundError } from './file-not-found.error';
+export { DisallowedFilepathError } from './disallowed-filepath.error';
+export { InvalidModeError } from './invalid-mode.error';
+export { InvalidManagerError } from './invalid-manager.error';
+export { InvalidExecutionMetadata } from './invalid-execution-metadata.error';
