+++ packages/core/src/errors/invalid-manager.error.ts
@@ -1,0 +1,7 @@
+import { BinaryDataError } from './abstract/binary-data.error';
+
+export class InvalidManagerError extends BinaryDataError {
+	constructor(mode: string) {
+		super(`No binary data manager found for: ${mode}`);
+	}
+}
