+++ packages/core/src/index.ts
@@ -18,3 +18,4 @@ export * from './errors';
 export { ObjectStoreService } from './ObjectStore/ObjectStore.service.ee';
 export { BinaryData } from './BinaryData/types';
 export { isStoredMode as isValidNonDefaultMode } from './BinaryData/utils';
+export * from './ExecutionMetadata';
