+++ packages/workflow/src/errors/cli-subworkflow-operation.error.ts
@@ -1,0 +1,3 @@
+import { SubworkflowOperationError } from './subworkflow-operation.error';
+
+export class CliWorkflowOperationError extends SubworkflowOperationError {}
