+++ packages/workflow/src/errors/expression-extension.error.ts
@@ -1,0 +1,3 @@
+import { ExpressionError } from './expression.error';
+
+export class ExpressionExtensionError extends ExpressionError {}
