+++ packages/workflow/src/errors/index.ts
@@ -1,1 +1,15 @@
-export { ReportableError } from './reportable.error';
+export { ApplicationError } from './application.error';
+export { ExpressionError } from './expression.error';
+export { NodeApiError } from './node-api.error';
+export { NodeOperationError } from './node-operation.error';
+export { NodeSslError } from './node-ssl.error';
+export { WebhookTakenError } from './webhook-taken.error';
+export { WorkflowActivationError } from './workflow-activation.error';
+export { WorkflowDeactivationError } from './workflow-deactivation.error';
+export { WorkflowOperationError } from './workflow-operation.error';
+export { SubworkflowOperationError } from './subworkflow-operation.error';
+export { CliWorkflowOperationError } from './cli-subworkflow-operation.error';
+
+export { NodeError } from './abstract/node.error';
+export { ExecutionBaseError } from './abstract/execution-base.error';
+export { ExpressionExtensionError } from './expression-extension.error';
