+++ packages/workflow/src/errors/node-ssl.error.ts
@@ -1,0 +1,7 @@
+import { ExecutionBaseError } from './abstract/execution-base.error';
+
+export class NodeSslError extends ExecutionBaseError {
+	constructor(cause: Error) {
+		super("SSL Issue: consider using the 'Ignore SSL issues' option", { cause });
+	}
+}
