+++ packages/workflow/src/errors/workflow-deactivation.error.ts
@@ -1,0 +1,3 @@
+import { WorkflowActivationError } from './workflow-activation.error';
+
+export class WorkflowDeactivationError extends WorkflowActivationError {}
