+++ packages/workflow/src/index.ts
@@ -15,14 +15,10 @@ export * from './Interfaces';
 export * from './MessageEventBus';
 export * from './ExecutionStatus';
 export * from './Expression';
-export * from './ExpressionError';
-export * from './NodeErrors';
 export * from './NodeHelpers';
 export * from './RoutingNode';
 export * from './Workflow';
-export * from './WorkflowActivationError';
 export * from './WorkflowDataProxy';
-export * from './WorkflowErrors';
 export * from './WorkflowHooks';
 export * from './VersionedNodeType';
 export { LoggerProxy, NodeHelpers, ObservableObject, TelemetryHelpers };
